requirements.txt
setuptools~=69.0.3
pypdf2~=3.0.1
Pillow~=10.0.0
openpyxl~=3.1.2
xlrd~=1.2.0
python-docx~=0.8.11
